+++ src/ContaoManager/Plugin.php
@@ -12,9 +12,11 @@ use Contao\CoreBundle\ContaoCoreBundle;
 use Contao\ManagerPlugin\Bundle\BundlePluginInterface;
 use Contao\ManagerPlugin\Bundle\Config\BundleConfig;
 use Contao\ManagerPlugin\Bundle\Parser\ParserInterface;
+use Contao\ManagerPlugin\Config\ConfigPluginInterface;
 use HeimrichHannot\FilenameSanitizerBundle\HeimrichHannotContaoFilenameSanitizerBundle;
+use Symfony\Component\Config\Loader\LoaderInterface;
 
-class Plugin implements BundlePluginInterface
+class Plugin implements BundlePluginInterface, ConfigPluginInterface
 {
     /**
      * {@inheritdoc}
@@ -24,5 +26,13 @@ class Plugin implements BundlePluginInterface
         return [
             BundleConfig::create(HeimrichHannotContaoFilenameSanitizerBundle::class)->setLoadAfter([ContaoCoreBundle::class]),
         ];
+    }
+
+    public function registerContainerConfiguration(LoaderInterface $loader, array $managerConfig)
+    {
+        $loader->load('@HeimrichHannotContaoFilenameSanitizerBundle/Resources/config/datacontainers.yml');
+        $loader->load('@HeimrichHannotContaoFilenameSanitizerBundle/Resources/config/services.yml');
+        $loader->load('@HeimrichHannotContaoFilenameSanitizerBundle/Resources/config/listeners.yml');
+        $loader->load('@HeimrichHannotContaoFilenameSanitizerBundle/Resources/config/commands.yml');
     }
 }