+++ src/HeimrichHannotContaoFilenameSanitizerBundle.php
@@ -8,16 +8,8 @@
 
 namespace HeimrichHannot\FilenameSanitizerBundle;
 
-use HeimrichHannot\FilenameSanitizerBundle\DependencyInjection\FilenameSanitizerExtension;
 use Symfony\Component\HttpKernel\Bundle\Bundle;
 
 class HeimrichHannotContaoFilenameSanitizerBundle extends Bundle
 {
-    /**
-     * {@inheritdoc}
-     */
-    public function getContainerExtension()
-    {
-        return new FilenameSanitizerExtension();
-    }
 }