+++ src/Resources/contao/dca/tl_files.php
@@ -5,9 +5,11 @@ $dca = &$GLOBALS['TL_DCA']['tl_files'];
 /**
  * Callbacks
  */
-$dca['config']['onsubmit_callback']['sanitizeFilename'] = ['huh.filename_sanitizer.data_container.files_container', 'sanitizeFilename'];
+$dca['config']['onsubmit_callback']['sanitizeFilename'] = [\HeimrichHannot\FilenameSanitizerBundle\DataContainer\FilesContainer::class, 'sanitizeFilename'];
 
 /**
  * Fields
  */
-$dca['fields']['name']['save_callback']['contao-filename-sanitizer-bundle'] = ['huh.filename_sanitizer.data_container.files_container', 'sanitizeFilenameOnSave'];
+$dca['fields']['name']['save_callback']['contao-filename-sanitizer-bundle'] = [
+    \HeimrichHannot\FilenameSanitizerBundle\DataContainer\FilesContainer::class, 'sanitizeFilenameOnSave'
+];